Feature/rework docker scripts (#4)
* Add stashed changes forgotten in structure rework PR

* See if I messed up splitting include in build_docker and config

* Add start script for starting container and add parallel pipeline stages

* Cleanup Jenkinsfile a tiny little bit

+++ .docker/into.sh
@@ -1,17 +1,21 @@
 #!/bin/bash
 
-# base directory needs to be mounted at same location as on host for dod (docker out of docker)
-abs_base_dir=$(cd "$(dirname "${BASH_SOURCE[0]}")/../"; pwd)
+source $(dirname "${BASH_SOURCE[0]}")/build_docker.sh
 
-source $(dirname "${BASH_SOURCE[0]}")/include.sh
+if [[ $(docker ps | grep ${name}) ]]
+then
+    docker exec -it $name /bin/bash
+else
+    docker run -it --rm --name ${name} \
+        --volume="$HOME/.Xauthority:/root/.Xauthority:rw" \
+        --volume /etc/passwd:/etc/passwd:ro \
+        --volume /etc/group:/etc/group:ro \
+        --user $(id -u) \
+        --net=host \
+        --env="DISPLAY" \
+        -v $abs_root_dir:$abs_root_dir \
+        -w $abs_root_dir \
+        $dockername /bin/bash
+fi
 
-docker run -it --rm --name ${name} \
-    --volume="$HOME/.Xauthority:/root/.Xauthority:rw" \
-    --volume /etc/passwd:/etc/passwd:ro \
-    --volume /etc/group:/etc/group:ro \
-    --user $(id -u) \
-    --net=host \
-    --env="DISPLAY" \
-    -v $abs_base_dir:$abs_base_dir \
-    -w $abs_base_dir \
-    $dockername /bin/bash
+
